# BillboardBackend/src/Presentation/Billboard.API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Billboard.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/advertisingRequests/create-ads": {
      "post": {
        "tags": [
          "AdvertisingRequests"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json",
          "application/xml",
          "text/xml",
          "application/*+xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateAdsDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/advertisingRequests/advertising": {
      "get": {
        "tags": [
          "AdvertisingRequests"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/advertisingRequests/owner": {
      "get": {
        "tags": [
          "AdvertisingRequests"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/advertisingRequests/update-approval-status": {
      "put": {
        "tags": [
          "AdvertisingRequests"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json",
          "application/xml",
          "text/xml",
          "application/*+xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateAdvertisingRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/billboards": {
      "post": {
        "tags": [
          "Billboards"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json",
          "application/xml",
          "text/xml",
          "application/*+xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BillboardCreateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Billboards"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json",
          "application/xml",
          "text/xml",
          "application/*+xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BillboardUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/billboards/advertising": {
      "get": {
        "tags": [
          "Billboards"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/billboards/owner": {
      "get": {
        "tags": [
          "Billboards"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/billboards/create-ads": {
      "get": {
        "tags": [
          "Billboards"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/billboards/{id}": {
      "get": {
        "tags": [
          "Billboards"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Billboards"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json",
          "application/xml",
          "text/xml",
          "application/*+xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserRegisterDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/update-password": {
      "post": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json",
          "application/xml",
          "text/xml",
          "application/*+xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserPasswordUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/login-owner": {
      "post": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json",
          "application/xml",
          "text/xml",
          "application/*+xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserLoginDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/login-advertising": {
      "post": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json",
          "application/xml",
          "text/xml",
          "application/*+xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserLoginDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "BillboardCreateDto": {
      "type": "object",
      "properties": {
        "Code": {
          "type": "string"
        },
        "LocationTitle": {
          "type": "string"
        },
        "LocationCoordinate": {
          "type": "string"
        },
        "PhotoUrl": {
          "format": "byte",
          "type": "string"
        },
        "ExpireDateTime": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "BillboardUpdateDto": {
      "type": "object",
      "properties": {
        "Id": {
          "format": "int32",
          "type": "integer"
        },
        "Code": {
          "type": "string"
        },
        "LocationTitle": {
          "type": "string"
        },
        "LocationCoordinate": {
          "type": "string"
        },
        "PhotoUrl": {
          "format": "byte",
          "type": "string"
        },
        "ExpireDateTime": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "CreateAdsDto": {
      "type": "object",
      "properties": {
        "BillboardId": {
          "format": "int32",
          "type": "integer"
        },
        "RequestedDays": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "UpdateAdvertisingRequestDto": {
      "type": "object",
      "properties": {
        "Id": {
          "format": "int32",
          "type": "integer"
        },
        "IsApproval": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "UserLoginDto": {
      "type": "object",
      "properties": {
        "Email": {
          "type": "string"
        },
        "Password": {
          "type": "string"
        }
      }
    },
    "UserPasswordUpdateDto": {
      "type": "object",
      "properties": {
        "OldPassword": {
          "type": "string"
        },
        "NewPassword": {
          "type": "string"
        }
      }
    },
    "UserRegisterDto": {
      "type": "object",
      "properties": {
        "FullName": {
          "type": "string"
        },
        "Email": {
          "type": "string"
        },
        "PhoneNumber": {
          "type": "string"
        },
        "Password": {
          "type": "string"
        },
        "AccountType": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Place to add JWT with Bearer"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}